WIP
This is a WIP. But what I am doing here is I am creating an API to
connect with Spotify. I have made a storage container or class for each
of the data sets that Spotify returns. I have made them in such a way
that the data will not get loaded into it's container until it is used
once used it will be cached so it doesn't have to go through conversion
again. I am going to make it so that not a single item would have to be
remade by using the id of the item it will check to see if it already
exists and if it does to return the existing item instead of generating
a new one. This becomes useful for 2 reasons. 1: Speed, 2: Access to the
same items between modules. So for example if i delete a track in
module A it is also deleted in module B without the need to tell
module B directly that it has been deleted.