Adds router discovery via upnp
This is a multi step thing to develop the redirect url. it checks to see
what the external ip address is and also grabs the local ip address of
the computer. if they are not the same then we assume there is a router
present. so we send out an SSDP discovery packet to see if there is one.
and if so if no port has been passed it will randomly pick one. and from
there we tell the router to add a port forward to the ip of the computer
on either the port supplied or the one randomly generated. then we build
the redirect url either using the external ip or a host name if there
was a host name supplied with the port and pass that back along with one
of 3 messages 'Router Success', 'Router Failure', 'Router Unavailable'.

I have also provided a means to remove the port mapping if available on
program exit.